review_detector-API/requirements.txt
Django==4.2
djangorestframework==3.14.0
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
textblob==0.17.1
django-cors-header==
scikit-learn==1.2.2
